requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flasgger>=0.9.7
tensorflow-cpu>=2.13.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
